move ChannelType to tachyon.network
