+++ common/src/main/java/tachyon/network/ChannelType.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package tachyon.worker.netty;
+package tachyon.network;
 
 import io.netty.channel.epoll.Epoll;
